+++ mathematics/affine_transformation.py
@@ -6,7 +6,7 @@ from math import radians
 
 def make_rotation_matrix(degrees, homo=True):
     axes = ["x", "y", "z"]
-    rot_mat = np.identity(3, np.float64)
+    rot_mat = np.identity(3, np.float32)
     for d_id, degree in enumerate(degrees):
         if degree != 0:
             rot_mat = rot_mat @ rot_mat_compute.from_euler(axes[d_id], degree, degrees=True).as_matrix()
@@ -23,13 +23,13 @@ def to_homogeneous_matrix(mat):
 
 
 def make_translation_matrix(moves):
-    res = np.identity(4, np.float64)
+    res = np.identity(4, np.float32)
     res[:3, 3] = moves
     return res
 
 
 def make_scale_matrix(scales):
-    res = np.identity(4, np.float64)
+    res = np.identity(4, np.float32)
     res[0, 0] = scales[0]
     res[1, 1] = scales[1]
     res[2, 2] = scales[2]
@@ -42,7 +42,7 @@ def make_transformation_matrix(transforms):
     :param transforms: {'scale': [2, 4, 2], 'rotation': [0, 90, 0]}
     :return:
     """
-    res = np.identity(4, np.float64)
+    res = np.identity(4, np.float32)
     if "position" in transforms:
         trans_mat = make_translation_matrix(transforms["position"])
         res = res @ trans_mat
